refactor(query): Extract filter methods into focused submodules and fix LIKE DP bug
Split the large filter_impl.rs (~2500 lines) into focused submodules:
- inspection.rs: Boolean queries about filter structure (has_*, is_*)
- relation.rs: Relation filter extraction and path analysis
- split.rs: Filter decomposition for query planning
- json_match.rs: JSON object/scalar matching for aggregate filters
- eval/operators.rs: Operator evaluation (comparison, equality, pattern matching)
- op.rs: FilterOp enum with parse/as_str/is_logical/is_array_element_op

Fixed DP bug in like_pattern_match where '%' wildcard used diagonal (prev)
instead of left neighbor (dp[j]) for zero-character match case.
