+++ crates/query/src/mapper/filter/eval/mod.rs
@@ -1,18 +1,8 @@
 //! Filter evaluation logic
 //!
-//! This module will contain extracted evaluation functionality:
-//! - matches(), eval_conditions()
-//! - eval_relation_conditions()
-//! - eval_alias_conditions()
-//! - eval_array_element_op()
-//! - eval_op(), values_equal()
-//!
-//! Currently these are methods on Filter in filter_impl.rs.
-//! These files are placeholders for future extraction.
+//! This module contains the core evaluation functionality for filters:
+//! - `operators`: Comparison, equality, pattern matching, array operators
 
-// Future submodules:
-// mod core;      // matches(), eval_conditions()
-// mod relation;  // eval_relation_conditions()
-// mod alias;     // eval_alias_conditions()
-// mod array;     // eval_array_element_op()
-// mod operators; // eval_op(), values_equal()
+mod operators;
+
+pub use operators::{eval_op, values_equal};
