+++ crates/query/src/mapper/filter/filter_impl.rs
@@ -2,149 +2,12 @@
 
 use std::collections::HashMap;
 
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 
+use super::eval::{eval_op, values_equal};
+use super::op::FilterOp;
 use crate::document::DocumentMapping;
 use crate::error::{QueryError, Result};
-
-/// Filter operators for condition matching
-/// Uses Go DefraDB naming conventions for compatibility
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum FilterOp {
-    /// Equal (_eq)
-    #[serde(rename = "_eq")]
-    Eq,
-    /// Not equal (_neq) - Go DefraDB naming
-    #[serde(rename = "_neq", alias = "_ne")]
-    Ne,
-    /// Greater than (_gt)
-    #[serde(rename = "_gt")]
-    Gt,
-    /// Greater than or equal (_geq) - Go DefraDB naming
-    #[serde(rename = "_geq", alias = "_gte")]
-    Gte,
-    /// Less than (_lt)
-    #[serde(rename = "_lt")]
-    Lt,
-    /// Less than or equal (_leq) - Go DefraDB naming
-    #[serde(rename = "_leq", alias = "_lte")]
-    Lte,
-    /// In array (_in)
-    #[serde(rename = "_in")]
-    In,
-    /// Not in array (_nin)
-    #[serde(rename = "_nin")]
-    Nin,
-    /// Pattern match (_like)
-    #[serde(rename = "_like")]
-    Like,
-    /// Negated pattern match (_nlike)
-    #[serde(rename = "_nlike")]
-    Nlike,
-    /// Case-insensitive pattern match (_ilike)
-    #[serde(rename = "_ilike")]
-    Ilike,
-    /// Negated case-insensitive pattern match (_nilike)
-    #[serde(rename = "_nilike")]
-    Nilike,
-    /// Array contains value (_contains)
-    #[serde(rename = "_contains")]
-    Contains,
-    /// Array is contained in given array (_contained_in)
-    #[serde(rename = "_contained_in")]
-    ContainedIn,
-    /// Object/map has key (_has_key)
-    #[serde(rename = "_has_key")]
-    HasKey,
-    /// Logical AND (_and)
-    #[serde(rename = "_and")]
-    And,
-    /// Logical OR (_or)
-    #[serde(rename = "_or")]
-    Or,
-    /// Logical NOT (_not)
-    #[serde(rename = "_not")]
-    Not,
-    /// Any array element matches condition (_any)
-    #[serde(rename = "_any")]
-    Any,
-    /// All array elements match condition (_all)
-    #[serde(rename = "_all")]
-    All,
-    /// No array elements match condition (_none)
-    #[serde(rename = "_none")]
-    None,
-}
-
-impl FilterOp {
-    /// Parse a filter operator from string.
-    /// Accepts both Go DefraDB naming (_neq, _geq, _leq) and alternative naming (_ne, _gte, _lte).
-    pub fn parse(s: &str) -> Option<Self> {
-        match s {
-            "_eq" => Some(Self::Eq),
-            "_neq" | "_ne" => Some(Self::Ne),
-            "_gt" => Some(Self::Gt),
-            "_geq" | "_gte" | "_ge" => Some(Self::Gte),
-            "_lt" => Some(Self::Lt),
-            "_leq" | "_lte" | "_le" => Some(Self::Lte),
-            "_in" => Some(Self::In),
-            "_nin" => Some(Self::Nin),
-            "_like" => Some(Self::Like),
-            "_nlike" => Some(Self::Nlike),
-            "_ilike" => Some(Self::Ilike),
-            "_nilike" => Some(Self::Nilike),
-            "_contains" => Some(Self::Contains),
-            "_contained_in" => Some(Self::ContainedIn),
-            "_has_key" => Some(Self::HasKey),
-            "_and" => Some(Self::And),
-            "_or" => Some(Self::Or),
-            "_not" => Some(Self::Not),
-            "_any" => Some(Self::Any),
-            "_all" => Some(Self::All),
-            "_none" => Some(Self::None),
-            _ => None,
-        }
-    }
-
-    /// Get the string representation (uses Go DefraDB naming)
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Eq => "_eq",
-            Self::Ne => "_neq",
-            Self::Gt => "_gt",
-            Self::Gte => "_geq",
-            Self::Lt => "_lt",
-            Self::Lte => "_leq",
-            Self::In => "_in",
-            Self::Nin => "_nin",
-            Self::Like => "_like",
-            Self::Nlike => "_nlike",
-            Self::Ilike => "_ilike",
-            Self::Nilike => "_nilike",
-            Self::Contains => "_contains",
-            Self::ContainedIn => "_contained_in",
-            Self::HasKey => "_has_key",
-            Self::And => "_and",
-            Self::Or => "_or",
-            Self::Not => "_not",
-            Self::Any => "_any",
-            Self::All => "_all",
-            Self::None => "_none",
-        }
-    }
-
-    /// Check if this is a logical operator
-    pub fn is_logical(&self) -> bool {
-        matches!(self, Self::And | Self::Or | Self::Not)
-    }
-
-    /// Check if this is an array element operator
-    pub fn is_array_element_op(&self) -> bool {
-        matches!(self, Self::Any | Self::All | Self::None)
-    }
-}
 
 /// Filter condition containing parsed conditions.
 ///
@@ -172,17 +35,6 @@ impl Filter {
         self.conditions.is_empty()
     }
 
-    /// Check if the filter has field-based conditions (as opposed to operator-only conditions).
-    ///
-    /// Field-based: `{rating: {_gt: 4.8}}` - the key "rating" is a field name
-    /// Operator-only: `{_gt: 4.8}` - the key "_gt" is an operator
-    ///
-    /// This is used to determine whether to use `matches_json_object` (field-based)
-    /// or `matches_scalar_value` (operator-only) when filtering relation aggregates.
-    pub fn has_field_conditions(&self) -> bool {
-        self.conditions.keys().any(|k| FilterOp::parse(k).is_none())
-    }
-
     /// Combine this filter with another filter using AND logic.
     ///
     /// Returns a new filter where both conditions must be satisfied.
@@ -198,547 +50,6 @@ impl Filter {
     /// Get a reference to the conditions map
     pub fn conditions(&self) -> &HashMap<String, JsonValue> {
         &self.conditions
-    }
-
-    /// Get all field names referenced by this filter
-    pub fn referenced_fields(&self) -> Vec<String> {
-        let mut fields = Vec::new();
-        Self::collect_fields(&self.conditions, &mut fields);
-        fields
-    }
-
-    fn collect_fields(conditions: &HashMap<String, JsonValue>, fields: &mut Vec<String>) {
-        for (key, value) in conditions {
-            // Skip logical operators
-            if FilterOp::parse(key).is_some() {
-                match value {
-                    JsonValue::Array(arr) => {
-                        for item in arr {
-                            if let JsonValue::Object(obj) = item {
-                                let nested: HashMap<String, JsonValue> =
-                                    obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                Self::collect_fields(&nested, fields);
-                            }
-                        }
-                    }
-                    JsonValue::Object(obj) => {
-                        let nested: HashMap<String, JsonValue> =
-                            obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                        Self::collect_fields(&nested, fields);
-                    }
-                    _ => {}
-                }
-            } else {
-                // This is a field name
-                if !fields.contains(key) {
-                    fields.push(key.clone());
-                }
-            }
-        }
-    }
-
-    /// Check if this filter contains relation filters (filters through nested objects).
-    ///
-    /// Relation filters are conditions like `{author: {verified: {_eq: true}}}` where
-    /// the first level field is a relation and the nested object contains field conditions
-    /// rather than operators.
-    pub fn has_relation_filters(&self) -> bool {
-        Self::check_for_relation_filters(&self.conditions)
-    }
-
-    /// Get the names of relation fields referenced in this filter.
-    ///
-    /// Returns field names that have nested object conditions (not operators),
-    /// indicating they are relation filters.
-    pub fn relation_field_names(&self) -> Vec<String> {
-        let mut names = Vec::new();
-        Self::collect_relation_field_names(&self.conditions, &mut names);
-        names
-    }
-
-    fn collect_relation_field_names(
-        conditions: &HashMap<String, JsonValue>,
-        names: &mut Vec<String>,
-    ) {
-        for (key, value) in conditions {
-            // Check logical operators recursively
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or => {
-                        if let JsonValue::Array(arr) = value {
-                            for item in arr {
-                                if let JsonValue::Object(obj) = item {
-                                    let nested: HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    Self::collect_relation_field_names(&nested, names);
-                                }
-                            }
-                        }
-                    }
-                    FilterOp::Not => {
-                        if let JsonValue::Object(obj) = value {
-                            let nested: HashMap<String, JsonValue> =
-                                obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                            Self::collect_relation_field_names(&nested, names);
-                        }
-                    }
-                    _ => {}
-                }
-            } else if key == "_alias" {
-                // _alias is a special filter directive, not a relation field
-                continue;
-            } else if let JsonValue::Object(obj) = value {
-                // This is a field condition - check if it contains operators or nested fields
-                // If any key in the object is NOT an operator, it's a relation filter
-                let is_relation = obj.keys().any(|k| FilterOp::parse(k).is_none());
-                if is_relation && !names.contains(key) {
-                    names.push(key.clone());
-                }
-            }
-        }
-    }
-
-    /// Get all relation filter conditions.
-    ///
-    /// Returns an iterator over (relation_name, nested_filter) pairs for each relation
-    /// referenced in the filter.
-    ///
-    /// For a filter like `{author: {verified: {_eq: true}}, rating: {_gt: 4}}`:
-    /// - Returns: [("author", Filter({verified: {_eq: true}}))]
-    /// - "rating" is not a relation filter (it has operators, not nested fields)
-    pub fn relation_conditions(&self) -> Vec<(String, Filter)> {
-        let mut result = Vec::new();
-        for name in self.relation_field_names() {
-            if let Some(filter) = self.extract_relation_filter(&name) {
-                result.push((name, filter));
-            }
-        }
-        result
-    }
-
-    fn check_for_relation_filters(conditions: &HashMap<String, JsonValue>) -> bool {
-        for (key, value) in conditions {
-            // Check logical operators recursively
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or => {
-                        if let JsonValue::Array(arr) = value {
-                            for item in arr {
-                                if let JsonValue::Object(obj) = item {
-                                    let nested: HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    if Self::check_for_relation_filters(&nested) {
-                                        return true;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FilterOp::Not => {
-                        if let JsonValue::Object(obj) = value {
-                            let nested: HashMap<String, JsonValue> =
-                                obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                            if Self::check_for_relation_filters(&nested) {
-                                return true;
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            } else if key == "_alias" {
-                // _alias is a special filter directive, not a relation filter
-                continue;
-            } else if let JsonValue::Object(obj) = value {
-                // This is a field condition - check if it contains operators or nested fields
-                // If any key in the object is NOT an operator, it's a relation filter
-                for nested_key in obj.keys() {
-                    if FilterOp::parse(nested_key).is_none() {
-                        // This is a nested field name, not an operator - it's a relation filter
-                        return true;
-                    }
-                }
-            }
-        }
-        false
-    }
-
-    /// Get all multi-level relation paths in this filter.
-    ///
-    /// For a filter like `{author: {published: {rating: {_eq: 4.9}}}}`, this returns
-    /// `[["author", "published"]]` - the path through relations (not including the leaf scalar).
-    ///
-    /// This is used to detect filters that require nested joins beyond a single level.
-    pub fn get_multi_level_relation_paths(&self) -> Vec<Vec<String>> {
-        let mut paths = Vec::new();
-        Self::collect_relation_paths(&self.conditions, &mut Vec::new(), &mut paths);
-        // Only return paths with more than one element (multi-level)
-        paths.into_iter().filter(|p| p.len() > 1).collect()
-    }
-
-    /// Recursively collect relation paths from filter conditions.
-    ///
-    /// A relation path is a sequence of nested field names before reaching a scalar condition.
-    /// For `{author: {published: {rating: {_eq: 4.9}}}}`:
-    /// - "author" is a relation (its value has non-operator keys)
-    /// - "published" is a relation (its value has non-operator keys)
-    /// - "rating" is a scalar (its value only has operator keys like "_eq")
-    /// So the path is ["author", "published"].
-    ///
-    /// This function only saves the deepest path (the full chain of relations).
-    fn collect_relation_paths(
-        conditions: &HashMap<String, JsonValue>,
-        current_path: &mut Vec<String>,
-        paths: &mut Vec<Vec<String>>,
-    ) {
-        for (key, value) in conditions {
-            // Skip logical operators at this level
-            if FilterOp::parse(key).is_some() {
-                // For _and/_or/_not, recurse into their contents
-                if let Some(op) = FilterOp::parse(key) {
-                    match op {
-                        FilterOp::And | FilterOp::Or => {
-                            if let JsonValue::Array(arr) = value {
-                                for item in arr {
-                                    if let JsonValue::Object(obj) = item {
-                                        let nested: HashMap<String, JsonValue> = obj
-                                            .iter()
-                                            .map(|(k, v)| (k.clone(), v.clone()))
-                                            .collect();
-                                        Self::collect_relation_paths(&nested, current_path, paths);
-                                    }
-                                }
-                            }
-                        }
-                        FilterOp::Not => {
-                            if let JsonValue::Object(obj) = value {
-                                let nested: HashMap<String, JsonValue> =
-                                    obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                Self::collect_relation_paths(&nested, current_path, paths);
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-                continue;
-            }
-
-            // _alias is a special filter directive, not a relation path
-            if key == "_alias" {
-                continue;
-            }
-
-            // Check if this key maps to a relation (nested object with non-operator keys)
-            if let JsonValue::Object(obj) = value {
-                let has_non_operator_keys = obj.keys().any(|k| FilterOp::parse(k).is_none());
-
-                if has_non_operator_keys {
-                    // This is a relation - add to path and recurse
-                    current_path.push(key.clone());
-
-                    let nested: HashMap<String, JsonValue> =
-                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-
-                    // Check if nested has any relations (non-operator keys that are objects with non-operator keys)
-                    let nested_has_relations = nested.iter().any(|(k, v)| {
-                        if FilterOp::parse(k).is_some() {
-                            return false;
-                        }
-                        if let JsonValue::Object(inner) = v {
-                            inner.keys().any(|ik| FilterOp::parse(ik).is_none())
-                        } else {
-                            false
-                        }
-                    });
-
-                    if nested_has_relations {
-                        // Continue recursing - there are more relations
-                        Self::collect_relation_paths(&nested, current_path, paths);
-                    } else {
-                        // This is the deepest level - save the path
-                        paths.push(current_path.clone());
-                    }
-
-                    current_path.pop();
-                }
-            }
-        }
-    }
-
-    /// Extract filter conditions at a specific relation path.
-    ///
-    /// For a filter like `{author: {published: {rating: {_eq: 4.9}}}}` and path `["author"]`,
-    /// this returns a Filter with conditions `{published: {rating: {_eq: 4.9}}}`.
-    ///
-    /// For path `["author", "published"]`, returns `{rating: {_eq: 4.9}}`.
-    pub fn extract_filter_at_path(&self, path: &[String]) -> Option<Filter> {
-        if path.is_empty() {
-            return Some(self.clone());
-        }
-
-        Self::extract_at_path_recursive(&self.conditions, path)
-    }
-
-    fn extract_at_path_recursive(
-        conditions: &HashMap<String, JsonValue>,
-        path: &[String],
-    ) -> Option<Filter> {
-        if path.is_empty() {
-            return Some(Filter::from_conditions(conditions.clone()));
-        }
-
-        let current_key = &path[0];
-        let remaining_path = &path[1..];
-
-        // Look for the key at this level
-        if let Some(value) = conditions.get(current_key) {
-            if let Some(obj) = value.as_object() {
-                let nested: HashMap<String, JsonValue> =
-                    obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-
-                if remaining_path.is_empty() {
-                    // We've reached the end of the path - return this level's conditions
-                    return Some(Filter::from_conditions(nested));
-                } else {
-                    // Continue down the path
-                    return Self::extract_at_path_recursive(&nested, remaining_path);
-                }
-            }
-        }
-
-        // Also check inside _and and _or blocks
-        for (key, value) in conditions {
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or => {
-                        if let Some(arr) = value.as_array() {
-                            for item in arr {
-                                if let Some(obj) = item.as_object() {
-                                    let nested: HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    if let Some(filter) =
-                                        Self::extract_at_path_recursive(&nested, path)
-                                    {
-                                        return Some(filter);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Check if this filter contains an `_alias` directive.
-    ///
-    /// `_alias` filters reference aliased fields/relations and must be evaluated after joins
-    /// because the alias may reference a relation whose data hasn't been joined yet.
-    pub fn has_alias_filter(&self) -> bool {
-        self.conditions.contains_key("_alias")
-    }
-
-    /// Check if this filter is "complex" - contains relation conditions inside logical operators.
-    ///
-    /// A filter is complex when `_and`, `_or`, or `_not` contains a mix of scalar and relation
-    /// conditions. Complex filters cannot be split and must be evaluated as a whole after
-    /// the join when the merged document is available.
-    ///
-    /// Examples:
-    /// - `{_and: [{rating: {_ge: 4.0}}, {author: {verified: {_eq: true}}}]}` → COMPLEX
-    /// - `{author: {verified: {_eq: true}}}` → NOT COMPLEX (relation at root, no logical wrapper)
-    /// - `{_and: [{rating: {_ge: 4.0}}, {age: {_gt: 25}}]}` → NOT COMPLEX (only scalars)
-    pub fn is_complex(&self) -> bool {
-        Self::check_for_complex_filters(&self.conditions)
-    }
-
-    fn check_for_complex_filters(conditions: &HashMap<String, JsonValue>) -> bool {
-        for (key, value) in conditions {
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or => {
-                        if let JsonValue::Array(arr) = value {
-                            // Check if this logical block contains any relation filters
-                            for item in arr {
-                                if let JsonValue::Object(obj) = item {
-                                    let nested: HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    // If any item in the logical block has a relation filter, it's complex
-                                    if Self::check_for_relation_filters(&nested) {
-                                        return true;
-                                    }
-                                    // Also check recursively for nested complex filters
-                                    if Self::check_for_complex_filters(&nested) {
-                                        return true;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    FilterOp::Not => {
-                        if let JsonValue::Object(obj) = value {
-                            let nested: HashMap<String, JsonValue> =
-                                obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                            if Self::check_for_relation_filters(&nested) {
-                                return true;
-                            }
-                            if Self::check_for_complex_filters(&nested) {
-                                return true;
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-        false
-    }
-
-    /// Split this filter into scalar filters and relation filters.
-    ///
-    /// Returns (scalar_filter, relation_filter) where:
-    /// - scalar_filter contains conditions on direct fields (no nested relation conditions)
-    /// - relation_filter contains conditions that traverse relations
-    ///
-    /// This is used to apply scalar filters before TypeJoin and relation filters after.
-    pub fn split_by_relation(&self) -> (Option<Filter>, Option<Filter>) {
-        let mut scalar_conditions = HashMap::new();
-        let mut relation_conditions = HashMap::new();
-
-        for (key, value) in &self.conditions {
-            // Logical operators need special handling - we can't easily split them
-            // For now, if they contain relation filters, put the whole condition in relation_filter
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or | FilterOp::Not => {
-                        // Check if this logical block contains relation filters
-                        let has_relation = match value {
-                            JsonValue::Array(arr) => arr.iter().any(|item| {
-                                if let JsonValue::Object(obj) = item {
-                                    let nested: HashMap<String, JsonValue> =
-                                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                    Self::check_for_relation_filters(&nested)
-                                } else {
-                                    false
-                                }
-                            }),
-                            JsonValue::Object(obj) => {
-                                let nested: HashMap<String, JsonValue> =
-                                    obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                                Self::check_for_relation_filters(&nested)
-                            }
-                            _ => false,
-                        };
-                        if has_relation {
-                            relation_conditions.insert(key.clone(), value.clone());
-                        } else {
-                            scalar_conditions.insert(key.clone(), value.clone());
-                        }
-                    }
-                    _ => {
-                        scalar_conditions.insert(key.clone(), value.clone());
-                    }
-                }
-            } else if key == "_alias" {
-                // _alias is a special filter directive evaluated by eval_conditions, not a relation
-                scalar_conditions.insert(key.clone(), value.clone());
-            } else if let JsonValue::Object(obj) = value {
-                // Field condition - check if it's a relation filter
-                let is_relation = obj.keys().any(|k| FilterOp::parse(k).is_none());
-                if is_relation {
-                    relation_conditions.insert(key.clone(), value.clone());
-                } else {
-                    scalar_conditions.insert(key.clone(), value.clone());
-                }
-            } else {
-                scalar_conditions.insert(key.clone(), value.clone());
-            }
-        }
-
-        let scalar = if scalar_conditions.is_empty() {
-            None
-        } else {
-            Some(Filter::from_conditions(scalar_conditions))
-        };
-
-        let relation = if relation_conditions.is_empty() {
-            None
-        } else {
-            Some(Filter::from_conditions(relation_conditions))
-        };
-
-        (scalar, relation)
-    }
-
-    /// Split the filter into non-alias and alias parts.
-    ///
-    /// Returns (non_alias_filter, alias_filter) where:
-    /// - non_alias_filter contains all conditions except `_alias`
-    /// - alias_filter contains only the `_alias` condition
-    ///
-    /// This is used to apply alias filters after aggregation in grouped queries,
-    /// since alias filters on aggregate fields can only be evaluated after the
-    /// aggregate values have been computed.
-    pub fn split_alias(&self) -> (Option<Filter>, Option<Filter>) {
-        let mut non_alias = HashMap::new();
-        let mut alias_only = HashMap::new();
-
-        for (key, value) in &self.conditions {
-            if key == "_alias" {
-                alias_only.insert(key.clone(), value.clone());
-            } else {
-                non_alias.insert(key.clone(), value.clone());
-            }
-        }
-
-        let non_alias_filter = if non_alias.is_empty() {
-            None
-        } else {
-            Some(Filter::from_conditions(non_alias))
-        };
-
-        let alias_filter = if alias_only.is_empty() {
-            None
-        } else {
-            Some(Filter::from_conditions(alias_only))
-        };
-
-        (non_alias_filter, alias_filter)
-    }
-
-    /// Strip _alias conditions from the filter that reference aggregate field names.
-    ///
-    /// This is used to prevent _alias conditions on computed aggregates from being
-    /// evaluated during plan execution (when aggregate values don't exist yet).
-    /// The stripped conditions should be applied in post-processing after aggregates are computed.
-    ///
-    /// Returns (filtered, has_aggregate_alias) where:
-    /// - filtered: Filter without _alias conditions referencing the given aggregate names
-    /// - has_aggregate_alias: true if any _alias conditions were referencing aggregates
-    pub fn strip_aggregate_alias_conditions(&self, aggregate_names: &[&str]) -> (Filter, bool) {
-        let mut new_conditions = HashMap::new();
-        let mut has_aggregate_alias = false;
-
-        for (key, value) in &self.conditions {
-            if key == "_alias" {
-                // Check if this _alias block references any aggregate names
-                if let Some(alias_obj) = value.as_object() {
-                    let refs_aggregate = alias_obj
-                        .keys()
-                        .any(|k| aggregate_names.contains(&k.as_str()));
-                    if refs_aggregate {
-                        has_aggregate_alias = true;
-                        // Skip this _alias condition - it will be applied in post-processing
-                        continue;
-                    }
-                }
-            }
-            new_conditions.insert(key.clone(), value.clone());
-        }
-
-        (Filter::from_conditions(new_conditions), has_aggregate_alias)
     }
 
     /// Evaluate the filter against document fields
@@ -853,7 +164,7 @@ impl Filter {
 
             // Handle null field conditions: {Name: null} is equivalent to {Name: {_eq: null}}
             if value.is_null() {
-                if !Self::values_equal(&field_value, &JsonValue::Null) {
+                if !values_equal(&field_value, &JsonValue::Null) {
                     return Ok(false);
                 }
                 continue;
@@ -918,7 +229,7 @@ impl Filter {
                         QueryError::invalid_filter(format!("unknown operator: {}", op_str))
                     })?;
 
-                    if !self.eval_op(&field_value, op, expected)? {
+                    if !eval_op(&field_value, op, expected)? {
                         return Ok(false);
                     }
                 }
@@ -1033,591 +344,13 @@ impl Filter {
                     let op = FilterOp::parse(op_str).ok_or_else(|| {
                         QueryError::invalid_filter(format!("unknown operator: {}", op_str))
                     })?;
-                    if !self.eval_op(&field_value, op, expected)? {
+                    if !eval_op(&field_value, op, expected)? {
                         return Ok(false);
                     }
                 }
             }
         }
         Ok(true)
-    }
-
-    fn eval_op(&self, actual: &JsonValue, op: FilterOp, expected: &JsonValue) -> Result<bool> {
-        match op {
-            FilterOp::Eq => Ok(Self::values_equal(actual, expected)),
-            FilterOp::Ne => Ok(!Self::values_equal(actual, expected)),
-            // Comparison operators: None (from null or NaN) returns false (Go DefraDB behavior)
-            FilterOp::Gt => self
-                .compare(actual, expected)
-                .map(|opt| opt.is_some_and(|ord| ord.is_gt())),
-            FilterOp::Gte => self
-                .compare(actual, expected)
-                .map(|opt| opt.is_some_and(|ord| ord.is_ge())),
-            FilterOp::Lt => self
-                .compare(actual, expected)
-                .map(|opt| opt.is_some_and(|ord| ord.is_lt())),
-            FilterOp::Lte => self
-                .compare(actual, expected)
-                .map(|opt| opt.is_some_and(|ord| ord.is_le())),
-            FilterOp::In => {
-                let arr = expected
-                    .as_array()
-                    .ok_or_else(|| QueryError::invalid_filter("_in requires array"))?;
-                Ok(arr.iter().any(|v| Self::values_equal(actual, v)))
-            }
-            FilterOp::Nin => {
-                let arr = expected
-                    .as_array()
-                    .ok_or_else(|| QueryError::invalid_filter("_nin requires array"))?;
-                Ok(!arr.iter().any(|v| Self::values_equal(actual, v)))
-            }
-            FilterOp::Like => self.like_match(actual, expected, false, false),
-            FilterOp::Nlike => self.like_match(actual, expected, true, false),
-            FilterOp::Ilike => self.like_match(actual, expected, false, true),
-            FilterOp::Nilike => self.like_match(actual, expected, true, true),
-            FilterOp::Contains => {
-                // Array field contains the expected value
-                // Null fields never match (standard database behavior)
-                if actual.is_null() {
-                    return Ok(false);
-                }
-                let arr = actual
-                    .as_array()
-                    .ok_or_else(|| QueryError::invalid_filter("_contains requires array field"))?;
-                Ok(arr.iter().any(|v| Self::values_equal(v, expected)))
-            }
-            FilterOp::ContainedIn => {
-                // All elements of actual array are in expected array (actual is subset of expected)
-                // Null fields never match (standard database behavior)
-                if actual.is_null() {
-                    return Ok(false);
-                }
-                let actual_arr = actual.as_array().ok_or_else(|| {
-                    QueryError::invalid_filter("_contained_in requires array field")
-                })?;
-                let expected_arr = expected.as_array().ok_or_else(|| {
-                    QueryError::invalid_filter("_contained_in requires array value")
-                })?;
-                Ok(actual_arr
-                    .iter()
-                    .all(|v| expected_arr.iter().any(|e| Self::values_equal(v, e))))
-            }
-            FilterOp::HasKey => {
-                // Object/map has the specified key
-                // Null fields never match (standard database behavior)
-                if actual.is_null() {
-                    return Ok(false);
-                }
-                let key = expected
-                    .as_str()
-                    .ok_or_else(|| QueryError::invalid_filter("_has_key requires string key"))?;
-                let obj = actual
-                    .as_object()
-                    .ok_or_else(|| QueryError::invalid_filter("_has_key requires object field"))?;
-                Ok(obj.contains_key(key))
-            }
-            FilterOp::Any => {
-                // Return true if ANY element matches the nested condition
-                // Null or non-array field → no match (matches Go DefraDB behavior for JSON fields)
-                let arr = match actual.as_array() {
-                    Some(a) => a,
-                    None => return Ok(false), // Non-array doesn't have any matching elements
-                };
-                // Empty array → no match (no elements to match)
-                if arr.is_empty() {
-                    return Ok(false);
-                }
-                // Expected is a nested condition object like {_gt: 70}
-                let nested_filter = expected
-                    .as_object()
-                    .ok_or_else(|| QueryError::invalid_filter("_any requires object condition"))?;
-                for elem in arr {
-                    if self.eval_conditions_on_value(elem, nested_filter)? {
-                        return Ok(true); // Found match
-                    }
-                }
-                Ok(false)
-            }
-            FilterOp::All => {
-                // Return true if ALL elements match the nested condition
-                // Non-array field → false (no elements, but not vacuous truth - matches Go behavior)
-                let arr = match actual.as_array() {
-                    Some(a) => a,
-                    None => return Ok(false), // Non-array fails _all
-                };
-                // Empty array → vacuous truth (all zero elements match)
-                if arr.is_empty() {
-                    return Ok(true);
-                }
-                let nested_filter = expected
-                    .as_object()
-                    .ok_or_else(|| QueryError::invalid_filter("_all requires object condition"))?;
-                for elem in arr {
-                    if !self.eval_conditions_on_value(elem, nested_filter)? {
-                        return Ok(false); // Found non-match
-                    }
-                }
-                Ok(true)
-            }
-            FilterOp::None => {
-                // Return true if NO elements match the nested condition
-                // Non-array field → false (not an array, so _none doesn't apply - Go behavior)
-                let arr = match actual.as_array() {
-                    Some(a) => a,
-                    None => return Ok(false), // Non-array excluded from _none results
-                };
-                // Empty array → true (no elements match)
-                if arr.is_empty() {
-                    return Ok(true);
-                }
-                let nested_filter = expected
-                    .as_object()
-                    .ok_or_else(|| QueryError::invalid_filter("_none requires object condition"))?;
-                for elem in arr {
-                    if self.eval_conditions_on_value(elem, nested_filter)? {
-                        return Ok(false); // Found match → fail
-                    }
-                }
-                Ok(true)
-            }
-            FilterOp::And | FilterOp::Or | FilterOp::Not => Err(QueryError::internal(
-                "logical ops should be handled at top level",
-            )),
-        }
-    }
-
-    fn values_equal(a: &JsonValue, b: &JsonValue) -> bool {
-        match (a, b) {
-            (JsonValue::Null, JsonValue::Null) => true,
-            (JsonValue::Bool(a), JsonValue::Bool(b)) => a == b,
-            (JsonValue::Number(a), JsonValue::Number(b)) => {
-                // Handle int/float comparison
-                if let (Some(a), Some(b)) = (a.as_i64(), b.as_i64()) {
-                    a == b
-                } else if let (Some(a), Some(b)) = (a.as_f64(), b.as_f64()) {
-                    (a - b).abs() < f64::EPSILON
-                } else {
-                    false
-                }
-            }
-            (JsonValue::String(a), JsonValue::String(b)) => {
-                // Try direct string comparison first
-                if a == b {
-                    return true;
-                }
-                // Try parsing as datetime values - stored values are in UTC format,
-                // filter values may have timezone offsets. Both represent the same time
-                // if they parse to the same UTC timestamp.
-                Self::datetimes_equal(a, b)
-            }
-            (JsonValue::Array(a), JsonValue::Array(b)) => {
-                a.len() == b.len() && a.iter().zip(b).all(|(a, b)| Self::values_equal(a, b))
-            }
-            (JsonValue::Object(a), JsonValue::Object(b)) => {
-                // Objects are equal if they have the same keys with equal values
-                if a.len() != b.len() {
-                    return false;
-                }
-                a.iter().all(|(key, val_a)| {
-                    b.get(key)
-                        .is_some_and(|val_b| Self::values_equal(val_a, val_b))
-                })
-            }
-            _ => false,
-        }
-    }
-
-    /// Try to compare two strings as datetime values.
-    /// Returns true if both can be parsed as RFC 3339 datetime strings and represent the same time.
-    fn datetimes_equal(a: &str, b: &str) -> bool {
-        let a_dt: Option<DateTime<Utc>> = a.parse().ok();
-        let b_dt: Option<DateTime<Utc>> = b.parse().ok();
-        match (a_dt, b_dt) {
-            (Some(a), Some(b)) => a == b,
-            _ => false, // If either isn't a valid datetime, they're not equal
-        }
-    }
-
-    /// Compare two values for ordering.
-    /// Implements Go DefraDB semantics:
-    /// - null vs null → Equal (null == null)
-    /// - null vs non-null → None (null document values are excluded from comparisons)
-    /// - non-null vs null → Greater (filter value is null, any non-null value is greater)
-    fn compare(&self, a: &JsonValue, b: &JsonValue) -> Result<Option<std::cmp::Ordering>> {
-        match (a, b) {
-            // Go DefraDB treats null as smallest value
-            // null vs null → Equal
-            (JsonValue::Null, JsonValue::Null) => Ok(Some(std::cmp::Ordering::Equal)),
-            // value vs null → Greater (any non-null value is greater than null)
-            (_, JsonValue::Null) => Ok(Some(std::cmp::Ordering::Greater)),
-            // null vs non-null → incomparable (Go DefraDB excludes null documents from comparisons)
-            (JsonValue::Null, _) => Ok(None),
-
-            // Number comparisons: support int/float coercion (Go's numbers.TryUpcast behavior)
-            (JsonValue::Number(a), JsonValue::Number(b)) => {
-                let a_val = a.as_f64().ok_or_else(|| {
-                    QueryError::invalid_filter(format!("number {} cannot be compared", a))
-                })?;
-                let b_val = b.as_f64().ok_or_else(|| {
-                    QueryError::invalid_filter(format!("number {} cannot be compared", b))
-                })?;
-                Ok(a_val.partial_cmp(&b_val)) // Returns None for NaN, which becomes false
-            }
-
-            // String comparisons - try datetime parsing first, then fall back to lexicographic
-            (JsonValue::String(a), JsonValue::String(b)) => {
-                // Try parsing as datetime values for proper temporal comparison
-                let a_dt: Option<DateTime<Utc>> = a.parse().ok();
-                let b_dt: Option<DateTime<Utc>> = b.parse().ok();
-                match (a_dt, b_dt) {
-                    (Some(a_time), Some(b_time)) => Ok(Some(a_time.cmp(&b_time))),
-                    _ => Ok(Some(a.cmp(b))), // Fall back to lexicographic comparison
-                }
-            }
-
-            // Type mismatch handling:
-            // - If filter value is a number but stored value isn't, return None (no match)
-            //   This is the "AllTypes" case where we have mixed-type JSON fields
-            // - If filter value is string, bool, object, or array, return an error
-            //   Go DefraDB only allows numeric comparisons with _gt/_ge/_lt/_le
-            (_, JsonValue::Number(_)) => {
-                // Filter value is number, but stored value type doesn't match
-                // Return no match instead of error
-                Ok(None)
-            }
-            // String filter values are NOT valid for comparison operators
-            (_, JsonValue::String(_)) => Err(QueryError::UnexpectedType {
-                property: "condition".to_string(),
-                actual: Self::go_type_name(b),
-            }),
-            // Error case: filter value (b) is an invalid type for comparison
-            // Use Go-compatible error format: "unexpected type. Property: condition, Actual: <type>"
-            _ => Err(QueryError::UnexpectedType {
-                property: "condition".to_string(),
-                actual: Self::go_type_name(b),
-            }),
-        }
-    }
-
-    /// Get Go-compatible type name for a JSON value
-    fn go_type_name(value: &JsonValue) -> String {
-        match value {
-            JsonValue::Null => "nil".to_string(),
-            JsonValue::Bool(_) => "bool".to_string(),
-            JsonValue::Number(_) => "float64".to_string(),
-            JsonValue::String(_) => "string".to_string(),
-            JsonValue::Array(_) => "[]interface {}".to_string(),
-            JsonValue::Object(_) => "map[string]interface {}".to_string(),
-        }
-    }
-
-    fn like_match(
-        &self,
-        actual: &JsonValue,
-        pattern: &JsonValue,
-        negate: bool,
-        case_insensitive: bool,
-    ) -> Result<bool> {
-        // Null fields never match (standard database behavior, matches Go DefraDB)
-        if actual.is_null() {
-            return Ok(negate);
-        }
-
-        let op_name = if case_insensitive { "_ilike" } else { "_like" };
-
-        // Non-string fields don't match _like (return false, not error)
-        // This matches Go DefraDB behavior for JSON fields with mixed types
-        let actual_str = match actual.as_str() {
-            Some(s) => s,
-            None => return Ok(negate), // Non-string doesn't match, so _like=false, _nlike=true
-        };
-        let pattern_str = pattern.as_str().ok_or_else(|| {
-            QueryError::invalid_filter(format!("{} requires string pattern", op_name))
-        })?;
-
-        // Apply case transformation if case-insensitive
-        let (actual_cmp, pattern_cmp): (std::borrow::Cow<str>, std::borrow::Cow<str>) =
-            if case_insensitive {
-                (
-                    actual_str.to_lowercase().into(),
-                    pattern_str.to_lowercase().into(),
-                )
-            } else {
-                (actual_str.into(), pattern_str.into())
-            };
-
-        // SQL LIKE pattern matching following Go DefraDB behavior:
-        // - '%' matches zero or more characters
-        // - '_' is treated as literal (matches Go behavior)
-        // - Supports arbitrary combinations of '%' wildcards
-        let matches = like_match(&actual_cmp, &pattern_cmp);
-
-        Ok(if negate { !matches } else { matches })
-    }
-
-    /// Evaluate the filter's conditions directly against a scalar JSON value.
-    ///
-    /// Used for inline array aggregate filters where each array element is tested
-    /// against operator conditions like `{_gt: 0}` or `{_and: [{_gt: -2}, {_lt: 2}]}`.
-    ///
-    /// Null values don't match comparison filters (they are excluded from aggregation).
-    pub fn matches_scalar_value(&self, value: &JsonValue) -> Result<bool> {
-        if self.conditions.is_empty() {
-            return Ok(true);
-        }
-        // Null scalar values don't pass comparison filters in inline array context
-        if value.is_null() {
-            return Ok(false);
-        }
-        for (key, expected) in &self.conditions {
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And => {
-                        let arr = expected
-                            .as_array()
-                            .ok_or_else(|| QueryError::invalid_filter("_and requires array"))?;
-                        for item in arr {
-                            let sub: HashMap<String, JsonValue> =
-                                serde_json::from_value(item.clone())
-                                    .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                            let f = Filter::from_conditions(sub);
-                            if !f.matches_scalar_value(value)? {
-                                return Ok(false);
-                            }
-                        }
-                    }
-                    FilterOp::Or => {
-                        let arr = expected
-                            .as_array()
-                            .ok_or_else(|| QueryError::invalid_filter("_or requires array"))?;
-                        let mut any_match = false;
-                        for item in arr {
-                            let sub: HashMap<String, JsonValue> =
-                                serde_json::from_value(item.clone())
-                                    .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                            let f = Filter::from_conditions(sub);
-                            if f.matches_scalar_value(value)? {
-                                any_match = true;
-                                break;
-                            }
-                        }
-                        if !any_match {
-                            return Ok(false);
-                        }
-                    }
-                    FilterOp::Not => {
-                        let sub: HashMap<String, JsonValue> =
-                            serde_json::from_value(expected.clone())
-                                .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                        let f = Filter::from_conditions(sub);
-                        if f.matches_scalar_value(value)? {
-                            return Ok(false);
-                        }
-                    }
-                    _ => {
-                        if !self.eval_op(value, op, expected)? {
-                            return Ok(false);
-                        }
-                    }
-                }
-            } else {
-                return Err(QueryError::invalid_filter(format!(
-                    "unknown operator in scalar filter: {}",
-                    key
-                )));
-            }
-        }
-        Ok(true)
-    }
-
-    /// Evaluate the filter against a JSON object (document).
-    ///
-    /// Used for relation aggregate filters where each related document is tested
-    /// against field-based conditions like `{rating: {_gt: 4.8}}`.
-    ///
-    /// Unlike `matches_scalar_value` which only handles operator conditions,
-    /// this method handles:
-    /// - Field-based conditions: `{rating: {_gt: 4.8}}` extracts `rating` from the object
-    /// - Operator-only conditions: `{_gt: 4.8}` (falls back to matches_scalar_value)
-    /// - Compound conditions: `{_and: [...]}`, `{_or: [...]}`, `{_not: {...}}`
-    pub fn matches_json_object(&self, obj: &JsonValue) -> Result<bool> {
-        if self.conditions.is_empty() {
-            return Ok(true);
-        }
-
-        for (key, expected) in &self.conditions {
-            if let Some(op) = FilterOp::parse(key) {
-                // Operator condition
-                match op {
-                    FilterOp::And => {
-                        let arr = expected
-                            .as_array()
-                            .ok_or_else(|| QueryError::invalid_filter("_and requires array"))?;
-                        for item in arr {
-                            let sub: HashMap<String, JsonValue> =
-                                serde_json::from_value(item.clone())
-                                    .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                            let f = Filter::from_conditions(sub);
-                            if !f.matches_json_object(obj)? {
-                                return Ok(false);
-                            }
-                        }
-                    }
-                    FilterOp::Or => {
-                        let arr = expected
-                            .as_array()
-                            .ok_or_else(|| QueryError::invalid_filter("_or requires array"))?;
-                        let mut any_match = false;
-                        for item in arr {
-                            let sub: HashMap<String, JsonValue> =
-                                serde_json::from_value(item.clone())
-                                    .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                            let f = Filter::from_conditions(sub);
-                            if f.matches_json_object(obj)? {
-                                any_match = true;
-                                break;
-                            }
-                        }
-                        if !any_match {
-                            return Ok(false);
-                        }
-                    }
-                    FilterOp::Not => {
-                        let sub: HashMap<String, JsonValue> =
-                            serde_json::from_value(expected.clone())
-                                .map_err(|e| QueryError::invalid_filter(e.to_string()))?;
-                        let f = Filter::from_conditions(sub);
-                        if f.matches_json_object(obj)? {
-                            return Ok(false);
-                        }
-                    }
-                    _ => {
-                        // Direct operator on the object itself - use scalar matching
-                        if !self.eval_op(obj, op, expected)? {
-                            return Ok(false);
-                        }
-                    }
-                }
-            } else {
-                // Field-based condition: key is a field name
-                let field_value = obj
-                    .as_object()
-                    .and_then(|o| o.get(key))
-                    .unwrap_or(&JsonValue::Null);
-
-                // expected should be an object with operator conditions
-                if let Some(conditions_obj) = expected.as_object() {
-                    // Check if the nested conditions are operators or field-based
-                    // If they're operators, use matches_scalar_value on the field value
-                    // If they're field-based, recursively use matches_json_object
-                    let has_field_conditions =
-                        conditions_obj.keys().any(|k| FilterOp::parse(k).is_none());
-                    let sub_conditions: HashMap<String, JsonValue> = conditions_obj
-                        .iter()
-                        .map(|(k, v)| (k.clone(), v.clone()))
-                        .collect();
-                    let f = Filter::from_conditions(sub_conditions);
-                    if has_field_conditions {
-                        // Nested field access - recursively match
-                        if !f.matches_json_object(field_value)? {
-                            return Ok(false);
-                        }
-                    } else {
-                        // Operator conditions on the field value
-                        if !f.matches_scalar_value(field_value)? {
-                            return Ok(false);
-                        }
-                    }
-                } else {
-                    // Direct equality check: {field: value}
-                    if field_value != expected {
-                        return Ok(false);
-                    }
-                }
-            }
-        }
-        Ok(true)
-    }
-
-    /// Evaluate a filter condition against a single JSON value (used by array element operators).
-    ///
-    /// For example, when evaluating `{_gt: 70}` against `85`, this method checks if 85 > 70.
-    fn eval_conditions_on_value(
-        &self,
-        value: &JsonValue,
-        conditions: &serde_json::Map<String, JsonValue>,
-    ) -> Result<bool> {
-        for (op_str, expected) in conditions {
-            let op = FilterOp::parse(op_str).ok_or_else(|| {
-                QueryError::invalid_filter(format!(
-                    "unknown operator in array condition: {}",
-                    op_str
-                ))
-            })?;
-            if !self.eval_op(value, op, expected)? {
-                return Ok(false);
-            }
-        }
-        Ok(true)
-    }
-
-    /// Extract the filter conditions for a specific relation field.
-    ///
-    /// For a filter like `{author: {verified: {_eq: true}}}`, calling
-    /// `extract_relation_filter("author")` returns a Filter with conditions
-    /// `{verified: {_eq: true}}`.
-    ///
-    /// Returns None if there's no filter condition for this relation field.
-    pub fn extract_relation_filter(&self, relation_field: &str) -> Option<Filter> {
-        // Check if this relation field has conditions at the top level
-        if let Some(value) = self.conditions.get(relation_field) {
-            if let Some(obj) = value.as_object() {
-                // Check if this is a nested filter (has non-operator keys)
-                let is_nested = obj.keys().any(|k| FilterOp::parse(k).is_none());
-                if is_nested {
-                    // Convert the nested object to a Filter
-                    let nested_conditions: HashMap<String, JsonValue> =
-                        obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect();
-                    return Some(Filter::from_conditions(nested_conditions));
-                }
-            }
-        }
-
-        // Also check inside _and and _or blocks
-        for (key, value) in &self.conditions {
-            if let Some(op) = FilterOp::parse(key) {
-                match op {
-                    FilterOp::And | FilterOp::Or => {
-                        if let Some(arr) = value.as_array() {
-                            for item in arr {
-                                if let Some(obj) = item.as_object() {
-                                    if let Some(rel_value) = obj.get(relation_field) {
-                                        if let Some(rel_obj) = rel_value.as_object() {
-                                            let is_nested = rel_obj
-                                                .keys()
-                                                .any(|k| FilterOp::parse(k).is_none());
-                                            if is_nested {
-                                                let nested_conditions: HashMap<String, JsonValue> =
-                                                    rel_obj
-                                                        .iter()
-                                                        .map(|(k, v)| (k.clone(), v.clone()))
-                                                        .collect();
-                                                return Some(Filter::from_conditions(
-                                                    nested_conditions,
-                                                ));
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        None
     }
 
     /// Evaluate alias-based filter conditions.
@@ -1701,7 +434,7 @@ impl Filter {
                 // Relation-style alias: {_alias: {books: {rating: {_gt: 4.8}}}}
                 // The alias points to a relation field (array or object)
                 if field_value.is_null() {
-                    // Null relation → no match
+                    // Null relation -> no match
                     return Ok(false);
                 } else if let Some(arr) = field_value.as_array() {
                     // One-to-many: existential semantics (any element matches)
@@ -1729,7 +462,7 @@ impl Filter {
                 // Scalar-style alias: {_alias: {myAge: {_gt: 20}}}
                 for (op_key, op_value) in ops {
                     if let Some(op) = FilterOp::parse(op_key) {
-                        if !self.eval_op(&field_value, op, op_value)? {
+                        if !eval_op(&field_value, op, op_value)? {
                             return Ok(false);
                         }
                     } else {
@@ -1743,47 +476,6 @@ impl Filter {
         }
         Ok(true)
     }
-}
-
-/// SQL LIKE pattern matching with `%` as wildcard for zero or more characters.
-/// `_` is treated as a literal character (matches Go DefraDB behavior).
-fn like_match(text: &str, pattern: &str) -> bool {
-    let text_bytes = text.as_bytes();
-    let pattern_bytes = pattern.as_bytes();
-    let t_len = text_bytes.len();
-    let p_len = pattern_bytes.len();
-
-    // dp[j] = true means text[0..i] matches pattern[0..j]
-    let mut dp = vec![false; p_len + 1];
-    dp[0] = true;
-
-    // Initialize: leading '%' can match empty string
-    for j in 0..p_len {
-        if pattern_bytes[j] == b'%' {
-            dp[j + 1] = dp[j];
-        } else {
-            break;
-        }
-    }
-
-    for i in 0..t_len {
-        let mut prev = dp[0];
-        dp[0] = false;
-        for j in 0..p_len {
-            let temp = dp[j + 1];
-            if pattern_bytes[j] == b'%' {
-                // '%' matches zero or more chars: either skip '%' (prev) or extend match (dp[j+1])
-                dp[j + 1] = prev || dp[j + 1];
-            } else if text_bytes[i] == pattern_bytes[j] {
-                dp[j + 1] = prev;
-            } else {
-                dp[j + 1] = false;
-            }
-            prev = temp;
-        }
-    }
-
-    dp[p_len]
 }
 
 #[cfg(test)]
@@ -2452,8 +1144,8 @@ mod tests {
     }
 
     #[test]
-    fn test_like_unsupported_complex_pattern() {
-        // Multiple % not at edges should error (except %contains%)
+    fn test_like_complex_pattern() {
+        // Multiple % should be handled by the DP algorithm
         let filter = Filter::from_conditions(HashMap::from([(
             "name".to_string(),
             json!({"_like": "%li%ce"}),
@@ -2461,7 +1153,8 @@ mod tests {
         let mapping = make_mapping();
         let fields = make_fields();
         let result = filter.matches(&fields, &mapping);
-        assert!(result.is_err());
+        // DP algorithm handles arbitrary % patterns
+        assert!(result.is_ok());
     }
 
     // =========================================================================
